+++ s3store/store.go
@@ -6,6 +6,7 @@ import (
 	"github.com/go-pg/pg/v9"
 	hclog "github.com/hashicorp/go-hclog"
         "jaeger-s3/config"
+        "jaeger-s3/config/types"
 
 	"github.com/jaegertracing/jaeger/plugin/storage/grpc/shared"
 	"github.com/jaegertracing/jaeger/storage/dependencystore"
@@ -23,8 +24,7 @@ type Store struct {
 	writer *Writer
 }
 
-
-func NewStore(conf *config.Configuration, logger hclog.Logger) (*Store, func() error, error) {
+func NewStore(conf *config.Configuration, cfg *types.Config, logger hclog.Logger) (*Store, func() error, error) {
 	db := pg.Connect(&pg.Options{
 		Addr:     conf.Host,
 		User:     conf.Username,
@@ -32,7 +32,7 @@ func NewStore(conf *config.Configuration, logger hclog.Logger) (*Store, func() e
 	}) 
 
 	reader := NewReader(db, logger)
-	writer := NewWriter(db, logger)
+	writer := NewWriter(db, cfg, logger)
 
 	store := &Store{
 		db:     db,
